populate 'start()' entrypoint of plugin & delete 'starters.py' / Starters class & cleanup bugs
- change to more generic 'configure_paths()' install step and call this prior to reset if hasattr = True for the component (module)
- entirely remove starters.py / Starters class and cleanup bugs
