+++ electrumsv-sdk/electrumsv_sdk/builtin_components/electrumsv/__init__.py
@@ -1,1 +1,1 @@
-from .electrumsv import install, start, stop, reset, status_check
+from .electrumsv import install, start, stop, reset, status_check, configure_paths
