+++ electrumsv-sdk/electrumsv_sdk/builtin_components/electrumx/__init__.py
@@ -1,1 +1,1 @@
-from .electrumx import install, start, stop, reset, status_check
+from .electrumx import install, start, stop, reset, status_check, configure_paths
