+++ electrumsv-sdk/electrumsv_sdk/builtin_components/node/node.py
@@ -1,13 +1,15 @@
 import logging
-import os
+from pathlib import Path
 
-from electrumsv_sdk.components import ComponentOptions
+from electrumsv_node import electrumsv_node
+
+from electrumsv_sdk.components import ComponentOptions, ComponentName, Component, ComponentState
+from electrumsv_sdk.utils import get_directory_name
 
 from .install import fetch_node
 
 
-MODULE_DIR = os.path.dirname(os.path.abspath(__file__))
-COMPONENT_NAME = os.path.basename(MODULE_DIR)
+COMPONENT_NAME = get_directory_name(__file__)
 logger = logging.getLogger(COMPONENT_NAME)
 
 
@@ -18,7 +20,7 @@ def install(app_state):
     if not repo == "":  # default
         logger.error("ignoring --repo flag for node - not applicable.")
 
-    # 1) configure_paths_and_datadir - (NOT APPLICABLE)  # Todo - need to add this
+    # 1) configure_paths - (NOT APPLICABLE)  # Todo - need to add this
     # 2) fetch (as needed) - (SEE BELOW)
     fetch_node(app_state)
     # 3) pip install (or npm install) packages/dependencies - (NOT APPLICABLE)
@@ -26,7 +28,27 @@ def install(app_state):
 
 
 def start(app_state):
-    pass
+    component_name = ComponentName.NODE
+    process_pid = electrumsv_node.start()
+    id = app_state.get_id(component_name)
+    logging_path = Path(electrumsv_node.DEFAULT_DATA_PATH)\
+        .joinpath("regtest").joinpath("bitcoind.log")
+
+    component = Component(id, process_pid, component_name, electrumsv_node.BITCOIND_PATH,
+        f"http://rpcuser:rpcpassword@127.0.0.1:18332", logging_path=logging_path,
+        metadata={"datadir": electrumsv_node.DEFAULT_DATA_PATH}
+    )
+    if not electrumsv_node.is_node_running():
+        component.component_state = ComponentState.Failed
+        logger.error("bitcoin daemon failed to start")
+    else:
+        component.component_state = ComponentState.Running
+        logger.debug("Bitcoin daemon online")
+
+    app_state.component_store.update_status_file(component)
+    app_state.status_monitor_client.update_status(component)
+
+    # process handle not returned because node is stopped via rpc
 
 
 def stop(app_state):
