+++ electrumsv-sdk/electrumsv_sdk/builtin_components/whatsonchain/whatsonchain.py
@@ -1,13 +1,14 @@
 import logging
-import os
+import sys
 
-from electrumsv_sdk.components import ComponentOptions
+from electrumsv_sdk.components import ComponentOptions, ComponentName, Component, ComponentState
+from electrumsv_sdk.utils import get_directory_name
 
 from .install import fetch_whatsonchain, generate_run_script_whatsonchain, packages_whatsonchain
+from .start import check_node_for_woc
 
 DEFAULT_PORT_WHATSONCHAIN = 3002
-MODULE_DIR = os.path.dirname(os.path.abspath(__file__))
-COMPONENT_NAME = os.path.basename(MODULE_DIR)
+COMPONENT_NAME = get_directory_name(__file__)
 logger = logging.getLogger(COMPONENT_NAME)
 
 
@@ -16,7 +17,7 @@ def install(app_state):
     if not repo == "":  # default
         logger.error("ignoring --repo flag for whatsonchain - not applicable.")
 
-    # 1) configure_paths_and_datadir (SEE BELOW)
+    # 1) configure_paths (SEE BELOW)
     app_state.woc_dir = app_state.depends_dir.joinpath("woc-explorer")
 
     # 2) fetch (as needed) (SEE BELOW)
@@ -30,7 +31,28 @@ def install(app_state):
 
 
 def start(app_state):
-    pass
+    component_name = ComponentName.WHATSONCHAIN
+    logger.debug(f"Starting whatsonchain daemon...")
+    if not check_node_for_woc():
+        sys.exit(1)
+
+    script_path = app_state.derive_shell_script_path(component_name)
+    process = app_state.spawn_process(script_path)
+    id = app_state.get_id(component_name)
+    component = Component(id, process.pid, component_name,
+        str(app_state.woc_dir), "http://127.0.0.1:3002")
+    is_running = app_state.is_component_running(component_name, component.status_endpoint,
+                                                4, 3, 1.0)
+    if not is_running:
+        component.component_state = ComponentState.Failed
+        logger.error("woc server failed to start")
+        sys.exit(1)
+    else:
+        component.component_state = ComponentState.Running
+        logger.debug("Whatsonchain server online")
+    app_state.component_store.update_status_file(component)
+    app_state.status_monitor_client.update_status(component)
+    return process
 
 
 def stop(app_state):
