--- electrumsv-sdk/electrumsv_sdk/starters.py
@@ -1,321 +1,0 @@
-import asyncio
-import logging
-import subprocess
-import sys
-import time
-import os
-from pathlib import Path
-
-import aiorpcx
-import requests
-from electrumsv_node import electrumsv_node
-
-from .utils import trace_pid, trace_processes_for_cmd
-from .status_monitor_client import StatusMonitorClient
-
-from .components import Component, ComponentName, ComponentState, ComponentStore, \
-    ComponentOptions
-
-logger = logging.getLogger("starters")
-MODULE_DIR = os.path.dirname(os.path.abspath(__file__))
-
-
-class ComponentLaunchFailedError(Exception):
-    pass
-
-
-class Starters:
-    def __init__(self, app_state):
-        self.app_state = app_state
-        self.component_store = ComponentStore(self.app_state)
-        self.status_monitor_client = StatusMonitorClient(self.app_state)
-        self.component_store = ComponentStore(self.app_state)
-
-    def spawn_process(self, command: str):
-        if self.app_state.start_options[ComponentOptions.BACKGROUND]:
-            return self.spawn_in_background(command)
-        else:
-            return self.spawn_in_new_console(command)
-
-    def run_command_current_shell(self, command: str):
-        subprocess.run(command, shell=True, check=True)
-
-    def run_command_background(self, command: str):
-        if sys.platform in ('linux', 'darwin'):
-            process_handle = subprocess.Popen(f"nohup {command} &", shell=True,
-                                       stdin=subprocess.PIPE, stdout=subprocess.PIPE,
-                                       stderr=subprocess.STDOUT)
-            process_handle.wait()
-            return process_handle
-        elif sys.platform == 'win32':
-            logger.info(
-                "Running as a background process (without a console window) is not supported "
-                "on windows, spawning in a new console window")
-            process_handle = subprocess.Popen(
-                f"{command}", creationflags=subprocess.CREATE_NEW_CONSOLE
-            )
-            return process_handle
-
-    def run_command_new_window(self, command: str):
-        if sys.platform in ('linux', 'darwin'):
-            # todo gnome-terminal part will not work cross-platform for spawning new terminals
-            process_handle = subprocess.Popen(f"gnome-terminal -- {command}", shell=True,
-                stdin=subprocess.PIPE, stdout=subprocess.PIPE, stderr=subprocess.STDOUT)
-            process_handle.wait()
-            return process_handle
-
-        elif sys.platform == 'win32':
-            process_handle = self.run_command_background(command)
-            return process_handle
-
-    def linux_trace_pid(self, command: str, num_processes_before: int):
-        num_processes_after = len(trace_processes_for_cmd(command))
-        if num_processes_before == num_processes_after:
-            raise ComponentLaunchFailedError()
-
-        process_handle = trace_pid(command)
-        return process_handle
-
-    def spawn_in_background(self, command):
-        """for long-running processes / servers - on linux there is a process id
-        tracing step because Popen returns a pid for a detached process (not the one we actually
-        want)"""
-        if sys.platform in ('linux', 'darwin'):
-            num_processes_before = len(trace_processes_for_cmd(command))
-            self.run_command_background(command)
-            time.sleep(1)  # allow brief time window for process to fail at startup
-
-            process_handle = self.linux_trace_pid(command, num_processes_before)
-            return process_handle
-
-        elif sys.platform == 'win32':
-            process_handle = self.run_command_background(command)
-            return process_handle
-
-    def spawn_in_new_console(self, command):
-        """for long-running processes / servers - on linux there is a process id tracing step
-        because Popen returns a pid for a detached process (not the one we actually want)"""
-        if sys.platform in ('linux', 'darwin'):
-            num_processes_before = len(trace_processes_for_cmd(command))
-            try:
-                self.run_command_new_window(command)
-            except ComponentLaunchFailedError:
-                logger.error(f"failed to launch long-running process: {command}. On linux cloud "
-                             f"servers try using the --background flag e.g. electrumsv-sdk start "
-                             f"--background node.")
-                raise
-            time.sleep(1)  # allow brief time window for process to fail at startup
-
-            process_handle = self.linux_trace_pid(command, num_processes_before)
-            return process_handle
-
-        elif sys.platform == 'win32':
-            process_handle = self.run_command_new_window(command)
-            return process_handle
-
-    async def is_electrumx_running(self):
-        for sleep_time in (1, 2, 3):
-            try:
-                logger.debug("Polling electrumx...")
-                async with aiorpcx.connect_rs(host="127.0.0.1", port=51001) as session:
-                    result = await session.send_request("server.version")
-                    if result[1] == "1.4":
-                        return True
-            except Exception as e:
-                pass
-
-            await asyncio.sleep(sleep_time)
-        return False
-
-    def is_component_running(self, component_name: ComponentName, status_endpoint: str, retries:
-            int=6, duration: float=1.0, timeout: float=0.5) -> bool:
-        for sleep_time in [duration] * retries:
-            logger.debug(f"Polling {component_name}...")
-            try:
-                result = requests.get(status_endpoint, timeout=timeout)
-                result.raise_for_status()
-                return True
-            except Exception as e:
-                pass
-
-            time.sleep(sleep_time)
-        return False
-
-    def start_node(self):
-        component_name = ComponentName.NODE
-        process_pid = electrumsv_node.start()
-        id = self.app_state.get_id(component_name)
-        logging_path = Path(electrumsv_node.DEFAULT_DATA_PATH)\
-            .joinpath("regtest").joinpath("bitcoind.log")
-
-        component = Component(id, process_pid, component_name, electrumsv_node.BITCOIND_PATH,
-            f"http://rpcuser:rpcpassword@127.0.0.1:18332", logging_path=logging_path,
-            metadata={"datadir": electrumsv_node.DEFAULT_DATA_PATH}
-        )
-        if not electrumsv_node.is_node_running():
-            component.component_state = ComponentState.Failed
-            logger.error("bitcoin daemon failed to start")
-        else:
-            component.component_state = ComponentState.Running
-            logger.debug("Bitcoin daemon online")
-
-        self.component_store.update_status_file(component)
-        self.status_monitor_client.update_status(component)
-
-        # process handle not returned because node is stopped via rpc
-
-    def start_electrumx(self):
-        component_name = ComponentName.ELECTRUMX
-        logger.debug(f"Starting RegTest electrumx server...")
-        script_path = self.component_store.derive_shell_script_path(ComponentName.ELECTRUMX)
-        process = self.spawn_process(script_path)
-        id = self.app_state.get_id(component_name)
-        component = Component(id, process.pid, ComponentName.ELECTRUMX,
-            location=str(self.app_state.electrumx_dir),
-            status_endpoint="http://127.0.0.1:51001",
-            metadata={"data_dir": str(self.app_state.electrumx_data_dir)},
-            logging_path=None,
-        )
-
-        is_running = asyncio.run(self.is_electrumx_running())
-        if not is_running:
-            component.component_state = ComponentState.Failed
-            logger.error("Electrumx server failed to start")
-        else:
-            component.component_state = ComponentState.Running
-            logger.debug("Electrumx online")
-        time.sleep(3)
-        self.component_store.update_status_file(component)
-        self.status_monitor_client.update_status(component)
-        return process
-
-    def init_electrumsv_wallet_dir(self):
-        os.makedirs(self.app_state.electrumsv_data_dir.joinpath("regtest/wallets"), exist_ok=True)
-
-    def esv_check_node_and_electrumx_running(self):
-        if not electrumsv_node.is_running():
-            logger.debug("Electrumsv in RegTest mode requires a bitcoin node to be running... "
-                         "failed to connect")
-            sys.exit()
-
-        is_running = asyncio.run(self.is_electrumx_running())
-        if not is_running:
-            logger.debug("Electrumsv in RegTest mode requires electrumx to be running... "
-                         "failed to connect")
-
-    def start_electrumsv(self, is_first_run=False):
-        """Todo - this currently uses ugly hacks with starting and stopping the ESV wallet
-         in order to:
-        1) generate the config files (so that it can be directly edited) - would be obviated by
-        fixing this: https://github.com/electrumsv/electrumsv/issues/111
-        2) newly created wallet doesn't seem to be fully useable until after stopping the daemon."""
-        logger.debug(f"Starting RegTest electrumsv daemon...")
-        component_name = ComponentName.ELECTRUMSV
-
-        # Option (1) Only using offline cli interface to electrumsv
-        if len(self.app_state.component_args) != 0:
-            if self.app_state.component_args[0] in ['create_wallet', 'create_account', '--help']:
-                return
-
-        # Option (2) Running daemon or gui proper
-        self.esv_check_node_and_electrumx_running()
-        self.init_electrumsv_wallet_dir()
-
-        script_path = self.component_store.derive_shell_script_path(component_name)
-        process = self.spawn_process(script_path)
-        if is_first_run:
-            self.app_state.resetters.reset_electrumsv_wallet()  # create first-time wallet
-
-            if sys.platform in ("linux", "darwin"):
-                subprocess.run(f"pkill -P {process.pid}", shell=True)
-            elif sys.platform == "win32":
-                subprocess.run(f"taskkill.exe /PID {process.pid} /T /F", check=True)
-            return self.start_electrumsv(is_first_run=False)
-
-        id = self.app_state.get_id(component_name)
-        logging_path = self.app_state.electrumsv_data_dir.joinpath("logs")
-        metadata = {"config": str(self.app_state.electrumsv_data_dir.joinpath("regtest/config")),
-                    "datadir": str(self.app_state.electrumsv_data_dir)}
-
-        component = Component(id, process.pid, component_name,
-            str(self.app_state.electrumsv_dir), "http://127.0.0.1:9999", metadata=metadata,
-            logging_path=logging_path)
-
-        is_running = self.is_component_running(component_name,
-                                               component.status_endpoint, 5, 2)
-        if not is_running:
-            component.component_state = ComponentState.Failed
-            logger.error("Electrumsv failed to start")
-            sys.exit(1)
-        else:
-            component.component_state = ComponentState.Running
-            logger.debug("Electrumsv online")
-
-        self.component_store.update_status_file(component)
-        self.status_monitor_client.update_status(component)
-        return process
-
-    def start_status_monitor(self):
-        component_name = ComponentName.STATUS_MONITOR
-        logger.debug(f"Starting status monitor daemon...")
-        try:
-            script_path = self.component_store.derive_shell_script_path(component_name)
-            process = self.spawn_process(script_path)
-        except ComponentLaunchFailedError:
-            log_path = self.app_state.status_monitor_logging_path
-            log_files = os.listdir(log_path)
-            log_files.sort(reverse=True)  # get latest log file at index 0
-            if len(log_files) != 0:
-                logger.debug(f"see {log_path.joinpath(log_files[0])} for details")
-            sys.exit(1)
-
-        id = self.app_state.get_id(component_name)
-        component = Component(id, process.pid, component_name,
-            str(self.app_state.status_monitor_dir), "http://127.0.0.1:5000/api/status/get_status")
-
-        is_running = self.is_component_running(component_name,
-                                               component.status_endpoint, 4, 0.5)
-        if not is_running:
-            component.component_state = ComponentState.Failed
-            logger.error("Status_monitor failed to start")
-        else:
-            component.component_state = ComponentState.Running
-            logger.debug("Status_monitor online")
-        self.component_store.update_status_file(component)
-        return process
-
-    def start_whatsonchain(self):
-        component_name = ComponentName.WHATSONCHAIN
-        logger.debug(f"Starting whatsonchain daemon...")
-        if not self.check_node_for_woc():
-            sys.exit(1)
-
-        script_path = self.component_store.derive_shell_script_path(component_name)
-        process = self.spawn_process(script_path)
-        id = self.app_state.get_id(component_name)
-        component = Component(id, process.pid, component_name,
-            str(self.app_state.woc_dir), "http://127.0.0.1:3002")
-        is_running = self.is_component_running(component_name, component.status_endpoint, 5, 3, 1.0)
-        if not is_running:
-            component.component_state = ComponentState.Failed
-            logger.error("woc server failed to start")
-            sys.exit(1)
-        else:
-            component.component_state = ComponentState.Running
-            logger.debug("Whatsonchain server online")
-        self.component_store.update_status_file(component)
-        self.status_monitor_client.update_status(component)
-        return process
-
-    def check_node_for_woc(self):
-        if not electrumsv_node.is_running():
-            return False
-
-        result = electrumsv_node.call_any("getinfo")
-        block_height = result.json()['result']['blocks']
-        if block_height == 0:
-            logger.error(f"Block height=0. "
-                f"The Whatsonchain explorer requires at least 1 block to be mined. Hint: try: "
-                f"'electrumsv-sdk node generate 1'")
-            return False
-        return True
